app/Agent/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.1
mysql-connector-python==8.0.24
nltk==3.6.2
protobuf==3.15.8
regex==2021.4.4
requests==2.25.1
six==1.15.0
tqdm==4.60.0
urllib3==1.26.4
